Add same tests for Swift and refactor
