+++ Examples/SwiftExampleApp/SwiftExampleAppUITests/CardPayment/CardPaymentUITests.swift
@@ -217,106 +217,6 @@ final class CardPaymentUITests: XCTestCase {
         XCTAssertFalse(newCard.waitForExistence(timeout: 5), "Unable to delete card")
     }
 
-    func testSuccessfulPaymentWithBillingDetails() {
-        app.launchArguments += ["-should_ask_for_billing_information", "true"]
-        app.launch()
-        app.textWithIdentifier(Selectors.FeatureList.payWithCard)?.tap()
-        app.fillWithStandardCardDetails()
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        app.fillBillingInfoDetails(email: Constants.BillingInfo.validEmail,
-                                   phone: Constants.BillingInfo.validMobile,
-                                   addressOne: Constants.BillingInfo.validAddress,
-                                   addressTwo: Constants.BillingInfo.validAddressTwo,
-                                   city: Constants.BillingInfo.validCity,
-                                   postCode: Constants.BillingInfo.validPostcode)
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        tapCompleteButton(app)
-        assertResultObject(app, "Payment", "AuthCode: ", "Success")
-        assertBillingInfo(app, Constants.BillingInfo.validCity, Constants.BillingInfo.validAddress, Constants.BillingInfo.validAddressTwo, Constants.BillingInfo.validPostcode)
-    }
-
-    func testUKPostCodeValidation() {
-        app.launchArguments += ["-should_ask_for_billing_information", "true"]
-        app.launch()
-        app.textWithIdentifier(Selectors.FeatureList.payWithCard)?.tap()
-        app.fillWithStandardCardDetails()
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        app.fillBillingInfoDetails(email: Constants.BillingInfo.validEmail,
-                                   phone: Constants.BillingInfo.validMobile,
-                                   addressOne: Constants.BillingInfo.validAddress,
-                                   addressTwo: Constants.BillingInfo.validAddressTwo,
-                                   city: Constants.BillingInfo.validCity,
-                                   postCode: Constants.BillingInfo.invalidPostcode)
-        app.mobileField?.tap()
-        XCTAssertEqual(app.fieldErrorLabel, Constants.BillingInfo.invalidPostcodeLabel)
-        XCTAssertFalse(app.cardDetailsSubmitButton!.isEnabled)
-    }
-
-    func testUSPostCodeValidation() throws {
-        guard #available(iOS 16, *) else {
-            throw XCTSkip("Skipping due to picker wheel selection being buggy on lower versions of iOS")
-        }
-        app.launchArguments += ["-should_ask_for_billing_information", "true"]
-        app.launch()
-        app.textWithIdentifier(Selectors.FeatureList.payWithCard)?.tap()
-        app.fillWithStandardCardDetails()
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        app.countryField?.tap()
-        app.pickerWheels.element(boundBy: 0).adjust(toPickerWheelValue: "United States")
-        app.administrativeDivisionField?.tap()
-        app.pickerWheels.element(boundBy: 0).adjust(toPickerWheelValue: "California")
-
-        app.fillBillingInfoDetails(email: Constants.BillingInfo.validEmail,
-                                   phone: Constants.BillingInfo.validMobile,
-                                   addressOne: Constants.BillingInfo.validAddress,
-                                   addressTwo: Constants.BillingInfo.validAddressTwo,
-                                   city: Constants.BillingInfo.validCity,
-                                   postCode: Constants.BillingInfo.invalidPostcode)
-        app.cityField?.tap()
-        XCTAssertEqual(app.fieldErrorLabel, Constants.BillingInfo.invalidZIPCodeLabel)
-        XCTAssertFalse(app.cardDetailsSubmitButton!.isEnabled)
-    }
-
-    func testCAPostCodeValidation() throws {
-        guard #available(iOS 16, *) else {
-            throw XCTSkip("Skipping due to picker wheel selection being buggy on lower versions of iOS")
-        }
-        app.launchArguments += ["-should_ask_for_billing_information", "true"]
-        app.launch()
-        app.textWithIdentifier(Selectors.FeatureList.payWithCard)?.tap()
-        app.fillWithStandardCardDetails()
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        app.countryField?.tap()
-        app.pickerWheels.element(boundBy: 0).adjust(toPickerWheelValue: "Canada")
-        app.administrativeDivisionField?.tap()
-        app.pickerWheels.element(boundBy: 0).adjust(toPickerWheelValue: "Ontario")
-        app.fillBillingInfoDetails(email: Constants.BillingInfo.validEmail,
-                                   phone: Constants.BillingInfo.validMobile,
-                                   addressOne: Constants.BillingInfo.validAddress,
-                                   addressTwo: Constants.BillingInfo.validAddressTwo,
-                                   city: Constants.BillingInfo.validCity,
-                                   postCode: Constants.BillingInfo.invalidPostcode)
-        app.cityField?.tap()
-        sleep(1)
-        XCTAssertEqual(app.fieldErrorLabel, Constants.BillingInfo.invalidPostcodeLabel)
-        XCTAssertFalse(app.cardDetailsSubmitButton!.isEnabled)
-    }
-
-    func testBillingFieldsInputValidation() {
-        app.launchArguments += ["-should_ask_for_billing_information", "true"]
-        app.launch()
-        app.textWithIdentifier(Selectors.FeatureList.payWithCard)?.tap()
-        app.fillWithStandardCardDetails()
-        XCTAssertTrue(app.cardDetailsSubmitButton!.isEnabled)
-        app.cardDetailsSubmitButton?.tap()
-        validateBillingInfoFields(app)
-    }
-
     func testStepUpPaymentTransaction() {
         app.launchArguments += ["-challenge_request_indicator", "noPreference", "-sca_exemption", "lowValue"]
         app.launch()
